Added 12 x 18, updated output format
I added 12 x 18 counting,  but defined it at 13 x 19 so that it wouldn't interfere with 11 x 17. I need to come up with a better way to do this in the future, but this works for now. 

I also updated the output format to only count the SqFt of the large format scans to match with how we bill. Per page for small format, per SqFt for large format.